feat(detect-targets): use exec-probe for glibc detection on Linux
Replace the version-banner string matching in detect/linux.rs with the
synthesized-ELF exec probe; the legacy loader-path detection moves to
detect/linux/fallback.rs and is only consulted when the probe is
inconclusive (noexec temp dir, seccomp, LSM denial), so restricted
environments degrade to the previous behaviour.

Fixes glibc detection on distros that rebrand the loader banner, e.g.
Gentoo (#1999), and removes the need for per-distro special cases
(gcompat, Ubuntu 20.04 exit-127) on the happy path. armv7 also no
longer depends on the empirical loader path list from #2386, as the
probe uses the real ld-linux-armhf.so.3 interpreter path.
